+++ src/main/java/net/geforcemods/securitycraft/misc/CustomDamageSources.java
@@ -14,6 +14,7 @@ public class CustomDamageSources {
 	public static final DamageSource FAKE_WATER = new DamageSource("securitycraft.fakeWater").setDamageBypassesArmor();
 	public static final DamageSource ELECTRICITY = new DamageSource("securitycraft.electricity").setDamageBypassesArmor();
 	public static final DamageSource INCORRECT_PASSCODE = new DamageSource("securitycraft.incorrectPasscode");
+	public static final DamageSource IN_REINFORCED_WALL = new DamageSource("securitycraft.inReinforcedWall").setDamageBypassesArmor().setDamageIsAbsolute();
 
 	private CustomDamageSources() {}